FW: Bugfixed UartReceiverStartLength_pullLastSentence function - proper array index: startPatternLength[readerIt].

+++ FW/Inc/user/utils.h
@@ -15,7 +15,7 @@
 
 #define ARRAY_LENGTH(x)	(sizeof(x) / sizeof((x)[0]))
 
-#define	U16_LSB(x)		(x & 0xFF)
-#define	U16_MSB(x)		((x >> 8) & 0xFF)
+#define	U16_LSB(x)		((uint8_t)(((uint16_t)x) & 0xFF))
+#define	U16_MSB(x)		((uint8_t)((((uint16_t)x) >> 8) & 0xFF))
 
 #endif /* USER_UTILS_H_ */
